+++ src/Middleware/DocBlockAnnotations.php
@@ -9,7 +9,6 @@ use JsonMapper\Enums\Visibility;
 use JsonMapper\JsonMapperInterface;
 use JsonMapper\ValueObjects\AnnotationMap;
 use JsonMapper\ValueObjects\PropertyMap;
-use JsonMapper\ValueObjects\PropertyType;
 use JsonMapper\Wrapper\ObjectWrapper;
 use Psr\SimpleCache\CacheInterface;
 
@@ -45,50 +44,51 @@ class DocBlockAnnotations extends AbstractMiddleware
 
         foreach ($properties as $property) {
             $name = $property->getName();
-            $propertyDetails = $this->derivePropertyDetailsFromReflectionProperty($property);
-
-            if (is_null($propertyDetails)) {
+            $docBlock = $property->getDocComment();
+            if ($docBlock === false) {
                 continue;
             }
 
-            $property = PropertyBuilder::new()
+            $annotations = self::parseDocBlockToAnnotationMap($docBlock);
+
+            if (! $annotations->hasVar()) {
+                continue;
+            }
+
+            $types = explode('|', $annotations->getVar());
+            $nullable = in_array('null', $types, true);
+            $types = array_filter($types, static function(string $type) {
+                return $type !== 'null';
+            });
+
+            $builder = PropertyBuilder::new()
                 ->setName($name)
-                ->setType($propertyDetails->getType())
-                ->setIsNullable($propertyDetails->isNullable())
-                ->setVisibility(Visibility::fromReflectionProperty($property))
-                ->setIsArray($propertyDetails->isArray())
-                ->build();
+                ->setIsNullable($nullable)
+                ->setVisibility(Visibility::fromReflectionProperty($property));
+
+            /* A union type that has one of its types defined as array is to complex to understand */
+            if (in_array('array', $types, true)) {
+                $property = $builder->addType('mixed', true)->build();
+                $intermediatePropertyMap->addProperty($property);
+                continue;
+            }
+
+            foreach ($types as $type) {
+                $type = trim($type);
+                $isArray = substr($type, -2) === '[]';
+                if ($isArray) {
+                    $type = substr($type, 0, -2);
+                }
+                $builder->addType($type, $isArray);
+            }
+
+            $property = $builder->build();
             $intermediatePropertyMap->addProperty($property);
         }
 
         $this->cache->set($object->getName(), $intermediatePropertyMap);
 
         return $intermediatePropertyMap;
-    }
-
-    private function derivePropertyDetailsFromReflectionProperty(\ReflectionProperty $property): ?PropertyType
-    {
-        $docBlock = $property->getDocComment();
-        if ($docBlock === false) {
-            return null;
-        }
-
-        $annotations = self::parseDocBlockToAnnotationMap($docBlock);
-
-        if (! $annotations->hasVar()) {
-            return null;
-        }
-
-        $type = $annotations->getVar();
-        $nullable = stripos('|' . $type . '|', '|null|') !== false;
-        $cleanedType = str_replace(['null|', '|null'], '', $type);
-
-        $isArray = substr($cleanedType, -2) === '[]';
-        if ($isArray) {
-            $cleanedType = substr($cleanedType, 0, -2);
-        }
-
-        return new PropertyType($cleanedType, $nullable, $isArray);
     }
 
     public static function parseDocBlockToAnnotationMap(string $docBlock): AnnotationMap